Populate tests with random values
