+++ Gemfile
@@ -3,14 +3,12 @@ git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
 ruby '3.0.0'
 
-gem 'bcrypt',           '~> 3.1'
-gem 'sprockets',        '~> 4.0'
-gem 'carrierwave',      '~> 2.2'
-gem 'cloudinary',       '~> 1.20'
-gem 'haml-rails',       '~> 2.0'
-gem 'rexml',            '~> 3.2', '>= 3.2.4'
-gem 'factory_bot',      '~> 4.8', '>= 4.8.2'
-gem 'database_cleaner', '~> 2.0', '>= 2.0.1'
+gem 'bcrypt',       '~> 3.1'
+gem 'sprockets',    '~> 4.0'
+gem 'carrierwave',  '~> 2.2'
+gem 'cloudinary',   '~> 1.20'
+gem 'haml-rails',   '~> 2.0'
+gem 'rexml',        '~> 3.2', '>= 3.2.4'
 
 gem 'rails',        '~> 6.1.3', '>= 6.1.3.1'
 gem 'pg',           '~> 1.1'
@@ -22,9 +20,11 @@ gem 'jbuilder',     '~> 2.7'
 gem 'bootsnap',     '>= 1.4.4', require: false
 
 group :development, :test do
-  gem 'byebug',       platforms: [:mri, :mingw, :x64_mingw]
-  gem 'dotenv-rails', '~> 2.7'
-  gem 'rspec-rails',  '~> 5.0'
+  gem 'byebug',           platforms: [:mri, :mingw, :x64_mingw]
+  gem 'dotenv-rails',     '~> 2.7'
+  gem 'rspec-rails',      '~> 5.0'
+  gem 'factory_bot',      '~> 4.8', '>= 4.8.2'
+  gem 'database_cleaner', '~> 2.0', '>= 2.0.1'
 end
 
 group :development do
